+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 1.0a4 (unreleased)
 ------------------
 
+- Implement validation using Cloudflare's AMP linter API endpoint.
+  [hvelarde]
+
 - Add support for the ``<amp-pixel>`` tag;
   see package documentation for more information on this new feature.
   [hvelarde]